+++ modelstore/storage/hosted.py
@@ -104,6 +104,11 @@ class HostedStorage(CloudStorage):
         }
         self._post("set-metadata", data)
 
+    def get_meta_data(self, domain: str, model_id: str) -> dict:
+        """ Returns a model's meta data """
+        # @TODO coming soon
+        raise NotImplementedError()
+
     def download(self, local_path: str, domain: str, model_id: str = None):
         """Downloads an artifacts archive for a given (domain, model_id) pair.
         If no model_id is given, it defaults to the latest model in that