+++ modelstore/models/lightgbm.py
@@ -14,6 +14,7 @@
 import json
 import os
 from functools import partial
+from typing import Any
 
 from modelstore.models.model_manager import ModelManager
 from modelstore.storage.storage import CloudStorage
@@ -61,6 +62,14 @@ class LightGbmManager(ModelManager):
         that are available
         """
         return kwargs["model"].params
+
+    def load(self, model_path: str, meta_data: dict) -> Any:
+        """
+        Loads a model, stored in model_path,
+        back into memory
+        """
+        # @TODO
+        raise NotImplementedError()
 
 
 def save_model(tmp_dir: str, model: "lgb.Booster") -> str: