+++ modelstore/models/transformers.py
@@ -13,6 +13,7 @@
 #    limitations under the License.
 import os
 from functools import partial
+from typing import Any
 
 from modelstore.models.model_manager import ModelManager
 from modelstore.storage.storage import CloudStorage
@@ -78,6 +79,14 @@ class TransformersManager(ModelManager):
         https://huggingface.co/transformers/main_classes/configuration.html#transformers.PretrainedConfig
         """
         return kwargs["config"].to_dict()
+
+    def load(self, model_path: str, meta_data: dict) -> Any:
+        """
+        Loads a model, stored in model_path,
+        back into memory
+        """
+        # @TODO
+        raise NotImplementedError()
 
 
 def _save_transformers(