+++ modelstore/models/xgboost.py
@@ -13,6 +13,7 @@
 #    limitations under the License.
 import os
 from functools import partial
+from typing import Any
 
 from modelstore.models.common import save_json
 from modelstore.models.model_manager import ModelManager
@@ -67,6 +68,14 @@ class XGBoostManager(ModelManager):
         that are available
         """
         return kwargs["model"].get_xgb_params()
+
+    def load(self, model_path: str, meta_data: dict) -> Any:
+        """
+        Loads a model, stored in model_path,
+        back into memory
+        """
+        # @TODO
+        raise NotImplementedError()
 
 
 def save_model(tmp_dir: str, model: "xgb.XGBModel") -> str: